+++ Source/Rendering/TrianglePrimitive.cpp
@@ -6,7 +6,7 @@ namespace greg
 {
 TrianglePrimitive::TrianglePrimitive()
 : vertices(std::vector<Vertex>(3)),
-  indices(std::vector<size_t>(3))
+  indices(std::vector<uint32_t>(3))
 {
     vertices[0] = Vertex(Vec2f(0.0f, -0.5f), Color(1.0f, 0.0f, 0.0f, 1.0f));
     vertices[1] = Vertex(Vec2f(0.5f, 0.5f), Color(0.0f, 1.0f, 0.0f, 1.0f));
@@ -22,7 +22,7 @@ const std::vector<greg::Vertex>& TrianglePrimitive::GetVertices() const
     return vertices;
 }
 
-const std::vector<size_t>& TrianglePrimitive::GetIndices() const
+const std::vector<uint32_t>& TrianglePrimitive::GetIndices() const
 {
     return indices;
 }
